+++ sdk/src/main/java/com/eyeson/sdk/webrtc/PeerConnectionClient.kt
@@ -91,7 +91,8 @@ internal class PeerConnectionClient(
     private val rootEglBase: EglBase,
     private val peerConnectionParameters: PeerConnectionParameters,
     private val events: PeerConnectionEvents,
-    private val dataChannelEvents: DataChannelEvents
+    private val dataChannelEvents: DataChannelEvents,
+    private val experimentalFeatureStereo: Boolean = false
 ) {
     private val pcObserver = PCObserver()
     private val sdpObserver = SDPObserver()
@@ -247,7 +248,6 @@ internal class PeerConnectionClient(
     val isVideoCallEnabled: Boolean
         get() = peerConnectionParameters.videoCallEnabled && videoCapturer != null
 
-    @Suppress("DEPRECATION")
     private fun createPeerConnectionFactoryInternal(options: PeerConnectionFactory.Options?) {
         isError = false
         if (peerConnectionParameters.tracing) {
@@ -330,12 +330,17 @@ internal class PeerConnectionClient(
                 reportError(errorMessage)
             }
         }
-        return JavaAudioDeviceModule.builder(appContext)
-            .setUseHardwareAcousticEchoCanceler(!peerConnectionParameters.disableBuiltInAEC)
-            .setUseHardwareNoiseSuppressor(!peerConnectionParameters.disableBuiltInNS)
-            .setAudioRecordErrorCallback(audioRecordErrorCallback)
-            .setAudioTrackErrorCallback(audioTrackErrorCallback)
-            .createAudioDeviceModule()
+        return JavaAudioDeviceModule.builder(appContext).apply {
+            if (experimentalFeatureStereo) {
+                setUseStereoInput(true)
+                setUseStereoOutput(true)
+            }
+            setUseHardwareAcousticEchoCanceler(!peerConnectionParameters.disableBuiltInAEC)
+            setUseHardwareNoiseSuppressor(!peerConnectionParameters.disableBuiltInNS)
+            setAudioRecordErrorCallback(audioRecordErrorCallback)
+            setAudioTrackErrorCallback(audioTrackErrorCallback)
+
+        }.createAudioDeviceModule()
     }
 
     private fun createMediaConstraintsInternal() {
@@ -1092,7 +1097,7 @@ internal class PeerConnectionClient(
             videoCodec = "VP8",
             videoCodecHwAcceleration = true,
             videoFlexfecEnabled = true,
-            audioStartBitrate = 200,
+            audioStartBitrate = 64,
             audioCodec = "",
             noAudioProcessing = false,
             aecDump = false,
@@ -1281,6 +1286,11 @@ internal class PeerConnectionClient(
                     false
                 )
             }
+
+            if (experimentalFeatureStereo) {
+                sdpDescription = setCodecStereo(sdpDescription, AUDIO_CODEC_OPUS, true)
+            }
+
             val sdp = SessionDescription(origSdp.type, "${sdpDescription.trim()}\r\n")
             localSdp = sdp
             executor.execute {
@@ -1544,6 +1554,46 @@ internal class PeerConnectionClient(
                 ?: return sdpDescription
             Logger.d("Change media description from: ${lines[mLineIndex]} to $newMLine")
             lines[mLineIndex] = newMLine
+            return joinString(listOf(*lines), "\r\n", true /* delimiterAtEnd */)
+        }
+
+        private fun setCodecStereo(
+            sdpDescription: String,
+            codec: String,
+            on: Boolean
+        ): String {
+            val lines = sdpDescription.split("\r\n".toRegex()).toTypedArray()
+            // a=rtpmap:<payload type> <encoding name>/<clock rate> [/<encoding parameters>]
+            val codecPattern = Pattern.compile("^a=rtpmap:(\\d+) $codec(/\\d+)+[\r]?$")
+            var id: String? = null
+            for (line in lines) {
+                val codecMatcher = codecPattern.matcher(line)
+                if (codecMatcher.matches()) {
+                    id = codecMatcher.group(1)
+                    break
+                }
+            }
+
+
+            val optionsPattern = Pattern.compile("^a=fmtp:$id (.+)\$")
+            val index = lines.indexOfFirst {
+                it.startsWith("a=fmtp:$id")
+            }
+
+            if (index >= 0) {
+                val optionsMatcher = optionsPattern.matcher(lines[index])
+                if (optionsMatcher.matches()) {
+                    val options =
+                        (optionsMatcher.group(1) ?: "").split(";".toRegex()).toMutableList()
+                    options.removeAll { it.contains("stereo") }
+                    options.add("stereo=${if (on) 1 else 0}")
+                    options.add("sprop-stereo=${if (on) 1 else 0}")
+
+                    lines[index] =
+                        "a=fmtp:$id ${joinString(options, ";", false /* delimiterAtEnd */)}"
+                }
+            }
+
             return joinString(listOf(*lines), "\r\n", true /* delimiterAtEnd */)
         }
     }